feat: add types for Albums
